ajout de EmployeDao, EntrepriseDao, UtilisateurDao, UtilisateurDaoImpl, Mysql (connection à la base de donnée) et légère modification pour Client et Entreprise

+++ src/codes/model/Entreprise.java
@@ -2,11 +2,10 @@ package codes.model;
 
 public class Entreprise extends Utilisateur{
 
-
     private String nom_entreprise;
-    private int siret;
+    private long siret;
 
-    public Entreprise(int id_utilisateur, String mdp, String email, String nom_entreprise, int siret){
+    public Entreprise(int id_utilisateur, String mdp, String email, String nom_entreprise, long siret){
         super(id_utilisateur, mdp, email);
         this.nom_entreprise = nom_entreprise;
         this.siret = siret;
@@ -20,11 +19,11 @@ public class Entreprise extends Utilisateur{
         this.nom_entreprise = nom_entreprise;
     }
 
-    public int getSiret() {
+    public long getSiret() {
         return siret;
     }
 
-    public void setSiret(int siret) {
+    public void setSiret(long siret) {
         this.siret = siret;
     }
 }
